requirements.txt
Django==3.2.5
djangorestframework==3.12.4
sqlalchemy
jsonify
pyodbc
pandas
django-pyodbc-azure
mssql-django